requirements.txt
pytest>=7.0.0
hvac>=1.2.0
requests>=2.28.0
pytest-html>=3.2.0
python-dotenv>=1.0.0
flake8>=7.0.0
black>=24.0.0
pytest-cov>=4.0.0
